django_project/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-utils-six==2.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
